feat: added resume method to set initial conditions to solution

+++ raven/models/base.py
@@ -105,7 +105,7 @@ class Raven:
         self.ostrich_exec = raven.ostrich_exec
         self._name = None
         self._defaults = {}
-        self.rvfiles = []
+        self.rvfiles = {}
 
         # Configuration file extensions + rvd for derived parameters.
         self._rvext = self._rvext + ('rvd',)
@@ -188,15 +188,7 @@ class Raven:
 
     @name.setter
     def name(self, x):
-        if self._name is None:
-            self._name = x
-        elif x != self._name:
-            raise UserWarning("Model configuration name changed.")
-        try:
-            if self.rvi.run_name is None:
-                self.rvi.run_name = x
-        except AttributeError:
-            pass
+        self._name = x
 
     @property
     def configuration(self):
@@ -225,7 +217,7 @@ class Raven:
             else:
                 if rvf.ext != 'txt':
                     setattr(self, 'name', rvf.stem)
-                self.rvfiles.append(rvf)
+                self.rvfiles[rvf.ext] = rvf
 
     def assign(self, key, value):
         """Assign parameter to rv object that has a key with the same name."""
@@ -256,7 +248,7 @@ class Raven:
         """Write configuration files to disk."""
         params = self.parameters
 
-        for rvf in self.rvfiles:
+        for rvf in self.rvfiles.values():
             p = self.exec_path if rvf.is_tpl else self.model_path
             if rvf.stem == 'OstRandomNumbers' and isinstance(self.txt, Ost) and self.txt.random_seed == "":
                 continue
@@ -273,23 +265,17 @@ class Raven:
            output/
 
         """
-        if self.exec_path.exists():
-            if overwrite:
+        if overwrite:
+            if self.model_path.exists():
                 shutil.rmtree(str(self.exec_path))
-            else:
-                raise IOError(
-                    "Directory already exists. Either set overwrite to `True` or create a new model instance.")
-
-        if self.final_path.exists():
-            if overwrite:
+            if self.final_path.exists():
                 shutil.rmtree(str(self.final_path))
-            else:
-                raise IOError(
-                    "Directory already exists. Either set overwrite to `True` or create a new model instance.")
 
         # Create general subdirectories
-        os.makedirs(str(self.exec_path))  # workdir/exec
-        os.makedirs(str(self.final_path))  # workdir/final
+        if not self.exec_path.exists():
+            os.makedirs(str(self.exec_path))  # workdir/exec
+        if not self.final_path.exists():
+            os.makedirs(str(self.final_path))  # workdir/final
 
     def setup_model_run(self, ts):
         """Create directory structure to store model input files, executable and output results.
@@ -318,10 +304,12 @@ class Raven:
 
         # Create symbolic link to input files
         for fn in ts:
-            os.symlink(str(fn), str(self.model_path / Path(fn).name))
+            if not (self.model_path / Path(fn).name).exists():
+                os.symlink(str(fn), str(self.model_path / Path(fn).name))
 
         # Create symbolic link to Raven executable
-        os.symlink(self.raven_exec, str(self.raven_cmd))
+        if not self.raven_cmd.exists():
+            os.symlink(self.raven_exec, str(self.raven_cmd))
 
         # Shell command to run the model
         if self.singularity:
@@ -434,6 +422,15 @@ class Raven:
         """.format(dir=self.cmd_path, err=err)
             print(msg)
             raise e
+
+    def resume(self):
+        """Set the initial state to the state at the end of the last run.
+
+        Note that the starting date of the next run should be identical to the end date of the previous run.
+        """
+        rvc = RVFile(self.outputs['solution'])
+        rvc.rename(self.name)
+        self.rvfiles['rvc'] = rvc
 
     def parse_results(self, path=None):
         """Store output files in the self.outputs dictionary."""
@@ -655,7 +652,7 @@ class Raven:
     def tags(self):
         """Return a list of tags within the templates."""
         out = []
-        for rvf in self.rvfiles:
+        for rvf in self.rvfiles.values():
             out.extend(rvf.tags)
 
         return out
